requirements.txt
brother-ql-inventree
numpy
Flask
pdf2image
